requirements.txt
python-dotenv
requests
watchdog
langchain-text-splitters
gitpython
PyYAML
